Address warnings on strict compile flags

+++ source/pbat/common/ArgSort.h
@@ -13,7 +13,7 @@ namespace common {
 template <std::integral TIndex, class FLess>
 std::vector<TIndex> ArgSort(auto n, FLess less)
 {
-    std::vector<TIndex> inds(n);
+    std::vector<TIndex> inds(static_cast<std::size_t>(n));
     std::iota(inds.begin(), inds.end(), TIndex(0));
     std::sort(inds.begin(), inds.end(), less);
     return inds;
